examiner_with_upload/requirements.txt
absl-py==2.0.0
altgraph==0.17.4
annotated-types==0.6.0
anyio==4.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
astor==0.8.1
asttokens==2.4.1
async-lru==2.0.4
attrs==23.1.0
Babel==2.13.1
beautifulsoup4==4.12.2
bing-image-downloader==1.1.2
bleach==1.5.0
blinker==1.7.0
blis==0.7.11
cachetools==5.3.2
camelot-py==0.11.0
catalogue==2.0.10
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
cloudpickle==3.0.0
colorama==0.4.6
comm==0.1.4
comtypes==1.2.0
confection==0.1.4
contourpy==1.1.1
cryptography==41.0.5
customtkinter==5.2.1
cv==1.0.0
cycler==0.12.1
cymem==2.0.8
darkdetect==0.8.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
Django==5.0
docutils==0.20.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
et-xmlfile==1.1.0
executing==2.0.1
fastjsonschema==2.18.1
Flask==3.0.0
fonttools==4.43.1
fqdn==1.5.1
gast==0.5.4
ghostscript==0.7
google-ai-generativelanguage==0.4.0
google-api-core==2.15.0
google-auth==2.25.2
google-cloud-documentai==2.21.1
google-generativeai==0.3.2
googleapis-common-protos==1.62.0
grpcio==1.60.0
grpcio-status==1.60.0
gTTS==2.5.0
gym==0.26.2
gym-notices==0.0.8
h11==0.14.0
html5lib==0.9999999
httpcore==1.0.2
httpx==0.26.0
idna==3.4
ipykernel==6.26.0
ipython==8.17.2
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
JPype1==1.5.0
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter-events==0.8.0
jupyter-lsp==2.2.0
jupyter_client==8.5.0
jupyter_core==5.5.0
jupyter_server==2.9.1
jupyter_server_terminals==0.4.4
jupyterlab==4.0.7
jupyterlab-pygments==0.2.2
jupyterlab_server==2.25.0
kivy-deps.angle==0.4.0
kivy-deps.glew==0.3.1
kivy-deps.gstreamer==0.3.3
kivy-deps.sdl2==0.7.0
kiwisolver==1.4.5
langcodes==3.3.0
lxml==5.0.0
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.8.1
matplotlib-inline==0.1.6
mistune==3.0.2
murmurhash==1.0.10
nbclient==0.8.0
nbconvert==7.10.0
nbformat==5.9.2
nest-asyncio==1.5.8
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.26.1
openai==1.6.1
opencv-python==4.8.1.78
openpyxl==3.1.2
overrides==7.4.0
packaging==23.2
pandas==2.1.2
pandocfilters==1.5.0
parso==0.8.3
pdfminer.six==20221105
pefile==2023.2.7
Pillow==10.1.0
platformdirs==3.11.0
plotly==5.18.0
preshed==3.0.9
prometheus-client==0.18.0
prompt-toolkit==3.0.39
proto-plus==1.23.0
protobuf==4.25.1
psutil==5.9.6
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAudio==0.2.14
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pygame==2.5.2
Pygments==2.16.1
pyinstaller==6.2.0
pyinstaller-hooks-contrib==2023.10
PyMuPDF==1.23.8
PyMuPDFb==1.23.7
pyparsing==3.1.1
pypdf==3.17.1
pypiwin32==223
pyproject-toml==0.0.10
python-dateutil==2.8.2
python-docx==1.1.0
python-json-logger==2.0.7
python-pptx==0.6.23
pyttsx3==2.90
pytz==2023.3.post1
pywin32==306
pywin32-ctypes==0.2.2
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.30.2
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.6
rsa==4.9
scikit-learn==1.3.2
scipy==1.11.3
seaborn==0.13.0
Send2Trash==1.8.2
setuptools==69.0.3
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
soupsieve==2.5
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.10.1
sqlparse==0.4.4
srsly==2.4.8
stack-data==0.6.3
tabula-py==2.9.0
tabulate==0.9.0
tenacity==8.2.3
tensorboard==1.8.0
termcolor==2.4.0
terminado==0.17.1
thinc==8.2.2
threadpoolctl==3.2.0
tinycss2==1.2.1
tk==0.1.0
toml==0.10.2
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
typer==0.9.0
types-python-dateutil==2.8.19.14
typing_extensions==4.9.0
tzdata==2023.3
uri-template==1.3.0
urllib3==2.0.7
wasabi==1.1.2
wcwidth==0.2.9
weasel==0.3.4
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
Werkzeug==3.0.1
wheel==0.42.0
XlsxWriter==3.1.9
gunicorn

